+++ src/compiler/codegen.h
@@ -4,10 +4,34 @@
 #include "symbols.h"
 #include "parser.h"
 
-int __flatcc_fb_codegen_common_c(fb_options_t *opts);
+typedef struct fb_output fb_output_t;
+
+struct fb_output {
+    /*
+     * Common namespace across files. May differ from namespace
+     * for consistent use of type names.
+     */
+    char nsc[FLATCC_NAMESPACE_MAX + 2];
+    char nscup[FLATCC_NAMESPACE_MAX + 2];
+
+    FILE *fp;
+    fb_schema_t *S;
+    fb_options_t *opts;
+    fb_scope_t *current_scope;
+    int indent;
+    int spacing;
+    int tmp_indent;
+};
+
+int __flatcc_fb_init_output_c(fb_output_t *out, fb_options_t *opts);
+#define fb_init_output_c __flatcc_fb_init_output_c
+void __flatcc_fb_end_output_c(fb_output_t *out);
+#define fb_end_output_c __flatcc_fb_end_output_c
+
+int __flatcc_fb_codegen_common_c(fb_output_t *out);
 #define fb_codegen_common_c __flatcc_fb_codegen_common_c
 
-int __flatcc_fb_codegen_c(fb_options_t *opts, fb_schema_t *S);
+int __flatcc_fb_codegen_c(fb_output_t *out, fb_schema_t *S);
 #define  fb_codegen_c __flatcc_fb_codegen_c
 
 void *__flatcc_fb_codegen_bfbs_to_buffer(fb_options_t *opts, fb_schema_t *S, void *buffer, size_t *size);