+++ src/compiler/codegen_c_sort.h
@@ -3,7 +3,7 @@
 
 #include "codegen_c.h"
 
-int __flatcc_gen_sort(output_t *out);
+int __flatcc_gen_sort(fb_output_t *out);
 #define gen_sort __flatcc_gen_sort
 
 #endif /* CODEGEN_SORT_C_H */