+++ src/compiler/flatcc.c
@@ -1,9 +1,12 @@
+#include <assert.h>
 #include "config.h"
 #include "parser.h"
 #include "semantics.h"
 #include "fileio.h"
 #include "codegen.h"
 #include "flatcc/flatcc.h"
+
+#define checkfree(s) if (s) { free(s); s = 0; }
 
 void flatcc_init_options(flatcc_options_t *opts)
 {
@@ -74,7 +77,7 @@ flatcc_context_t flatcc_create_context(flatcc_options_t *opts, const char *name,
     return P;
 }
 
-flatcc_context_t __flatcc_create_child_context(flatcc_options_t *opts, const char *name,
+static flatcc_context_t __flatcc_create_child_context(flatcc_options_t *opts, const char *name,
         fb_parser_t *P_parent)
 {
     fb_parser_t *P;
@@ -95,7 +98,7 @@ int flatcc_parse_buffer(flatcc_context_t ctx, const char *buf, size_t buflen)
 {
     fb_parser_t *P = ctx;
 
-    /* Currently includes cannot be handled by buffers, so they should fail. */
+    /* Currently includes cannot be handled by buffers, so they should done. */
     P->opts.disable_includes = 1;
     if ((size_t)buflen > P->opts.max_schema_size && P->opts.max_schema_size > 0) {
         fb_print_error(P, "input exceeds maximum allowed size\n");
@@ -159,6 +162,155 @@ static int __parse_include_file(fb_parser_t *P_parent, const char *filename)
     return 0;
 }
 
+/*
+ * The depends file format is a make rule:
+ *
+ * <outputfile> : <dep1-file> <dep2-file> ...
+ *
+ * like -MMD option for gcc/clang:
+ * lib.o.d generated with content:
+ *
+ * lib.o : header1.h header2.h
+ *
+ * We use a file name <basename>.depends for schema <basename>.fbs with content:
+ *
+ * <basename>_reader.h : <included-schema-1> ...
+ *
+ * The .d extension could mean the D language and we don't have sensible
+ * .o.d name because of multiple outputs, so .depends is better.
+ *
+ * (the above above is subject to the configuration of extensions).
+ *
+ * TODO:
+ * perhaps we should optionally add a dependency to the common reader
+ * and builder files when they are generated separately as they should in
+ * concurrent builds.
+ *
+ * TODO:
+ * 1. we should have a file for every output we produce (_builder.h * etc.)
+ * 2. reader might not even be in the output, e.g. verifier only.
+ * 3. multiple outputs doesn't work with ninja build 1.7.1, so just
+ *    use reader for now, and possible add an option for multiple
+ *    outputs later.
+ *
+ *  http://stackoverflow.com/questions/11855386/using-g-with-mmd-in-makefile-to-automatically-generate-dependencies
+ *  https://groups.google.com/forum/#!topic/ninja-build/j-2RfBIOd_8
+ *  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47485
+ *
+ *  Spaces in gnu make:
+ *  https://www.cmcrossroads.com/article/gnu-make-meets-file-names-spaces-them
+ *  See comments on gnu make handling of spaces.
+ *  http://clang.llvm.org/doxygen/DependencyFile_8cpp_source.html
+ */
+static int __flatcc_gen_depends_file(fb_parser_t *P)
+{
+    FILE *fp = 0;
+    const char *outpath, *basename; 
+    const char *depfile, *deproot, *depext;
+    const char *targetfile, *targetsuffix, *targetroot;
+    char *path = 0, *deppath = 0, *tmppath = 0, *targetpath = 0; 
+    int ret = -1;
+
+    /*
+     * The dependencies list is only correct for root files as it is a
+     * linear list. To deal with children, we would have to filter via
+     * the visible schema hash table, but we don't really need that.
+     */
+    assert(P->referer_path == 0);
+
+    outpath = P->opts.outpath ? P->opts.outpath : "";
+    basename = P->schema.basename;
+    targetfile = P->opts.gen_deptarget;
+
+
+    /* The following is mostly considering build tools generating
+     * a depfile as Ninja build would use it. It is a bit strict
+     * on path variations and currenlty doesn't accept multiple
+     * build products in a build rule (Ninja 1.7.1).
+     *
+     * Make depfile relative to cwd so the user can add output if
+     * needed, otherwise it is not possible, or difficult, to use a path given
+     * by a build tool, relative the cwd. If --depfile is not given,
+     * then -d is given or we would not be here. In that case we add an
+     * extension "<basename>.fbs.d" in the outpath.
+     *
+     * A general problem is that the outpath may be a build root dir or
+     * a current subdir for a custom build rule while the dep file
+     * content needs the same path every time, not just an equivalent
+     * path. For dependencies, we can rely on the input schema path.
+     * The input search paths may because confusion but we choose the
+     * discovered path relative to cwd consistently for each schema file
+     * encountered.
+     *
+     * The target file (<target>: <include1.fbs> <include2.fbs> ...)
+     * is tricky because it is not unique - but we can chose <schema>_reader.h
+     * or <schema>.bfbs prefixed with outpath. The user should choose an
+     * outpath relative to cwd or an absolute path depending on what the
+     * build system prefers. This may not be so easy in praxis, but what
+     * can we do?
+     *
+     * It is important to note the default target and the default
+     * depfile name is not just a convenience. Sometimes it is much
+     * simpler to use this version over an explicit path, sometimes
+     * perhaps not so much.
+     */
+
+    if (P->opts.gen_depfile) {
+        depfile = P->opts.gen_depfile;
+        deproot = "";
+        depext = "";
+    } else {
+        depfile = basename;
+        deproot = outpath;
+        depext = FLATCC_DEFAULT_DEP_EXT;
+    }
+    if (targetfile) {
+        targetsuffix = "";
+        targetroot = "";
+    } else {
+        targetsuffix = P->opts.bgen_bfbs 
+                ? FLATCC_DEFAULT_BIN_SCHEMA_EXT 
+                : FLATCC_DEFAULT_DEP_TARGET_SUFFIX;
+        targetfile = basename;
+        targetroot = outpath;
+    }
+
+    checkmem(path = fb_create_join_path(deproot, depfile, depext, 1));
+
+    checkmem(tmppath = fb_create_join_path(targetroot, targetfile, targetsuffix, 1));
+    /* Handle spaces in dependency file. */
+    checkmem((targetpath = fb_create_make_path(tmppath)));
+    checkfree(tmppath);
+
+    fp = fopen(path, "wb");
+    if (!fp) {
+        fb_print_error(P, "could not open dependency file for output: %s\n", path);
+        goto done;
+    }
+    fprintf(fp, "%s:", targetpath);
+
+    /* Don't depend on root schema. */
+    P = P->dependencies;
+    while (P) {
+        checkmem((deppath = fb_create_make_path(P->path)));
+        fprintf(fp, " %s", deppath);
+        P = P->dependencies;
+        checkfree(deppath);
+    }
+    fprintf(fp, "\n");
+    ret = 0;
+
+done:
+    checkfree(path);
+    checkfree(tmppath);
+    checkfree(targetpath);
+    checkfree(deppath);
+    if (fp) {
+        fclose(fp);
+    }
+    return ret;
+}
+
 int flatcc_parse_file(flatcc_context_t ctx, const char *filename)
 {
     fb_parser_t *P = ctx;
@@ -178,17 +330,19 @@ int flatcc_parse_file(flatcc_context_t ctx, const char *filename)
     path = 0;
     include_file = 0;
     ret = -1;
+    int is_root = !P->referer_path;
 
     /*
      * For root files, read file relative to working dir first. For
      * included files (`referer_path` set), first try include paths
      * in order, then path relative to including file.
      */
-    if (!P->referer_path) {
+    if (is_root) {
         if (!(buf = fb_read_file(filename, P->opts.max_schema_size, &size))) {
             if (size + P->schema.root_schema->total_source_size > P->opts.max_schema_size && P->opts.max_schema_size > 0) {
                 fb_print_error(P, "input exceeds maximum allowed size\n");
-                return -1;
+                ret = -1;
+                goto done;
             }
         } else {
             checkmem((path = fb_copy_path(filename)));
@@ -197,46 +351,52 @@ int flatcc_parse_file(flatcc_context_t ctx, const char *filename)
     for (i = 0; !buf && i < P->opts.inpath_count; ++i) {
         inpath = P->opts.inpaths[i];
         inpath_len = strlen(inpath);
-        checkmem((path = fb_create_join_path(inpath, inpath_len, filename, filename_len, "", 1)));
+        checkmem((path = fb_create_join_path_n(inpath, inpath_len, filename, filename_len, "", 1)));
         if (!(buf = fb_read_file(path, P->opts.max_schema_size, &size))) {
             free(path);
             path = 0;
             if (size > P->opts.max_schema_size && P->opts.max_schema_size > 0) {
                 fb_print_error(P, "input exceeds maximum allowed size\n");
-                return -1;
+                ret = -1;
+                goto done;
             }
         }
     }
-    if (!buf && P->referer_path) {
+    if (!buf && !is_root) {
         inpath = P->referer_path;
         inpath_len = fb_find_basename(inpath, strlen(inpath));
-        checkmem((path = fb_create_join_path(inpath, inpath_len, filename, filename_len, "", 1)));
+        checkmem((path = fb_create_join_path_n(inpath, inpath_len, filename, filename_len, "", 1)));
         if (!(buf = fb_read_file(path, P->opts.max_schema_size, &size))) {
             free(path);
             path = 0;
             if (size > P->opts.max_schema_size && P->opts.max_schema_size > 0) {
                 fb_print_error(P, "input exceeds maximum allowed size\n");
-                return -1;
+                ret = -1;
+                goto done;
             }
         }
     }
     if (!buf) {
         fb_print_error(P, "error reading included schema file: %s\n", filename);
-        return -1;
+        goto done;
     }
     P->schema.root_schema->total_source_size += size;
     P->path = path;
+    /* Parser owns path. */
+    path = 0;
     /*
      * Even if we do not have the recursive option set, we still
      * need to parse all include files to make sense of the current
      * file.
      */
     if (!(ret = fb_parse(P, buf, size, 1))) {
+        /* Parser owns buffer. */
+        buf = 0;
         inc = P->schema.includes;
         while (inc) {
             checkmem((include_file = fb_copy_path_n(inc->name.s.s, inc->name.s.len)));
             if (__parse_include_file(P, include_file)) {
-                return -1;
+                goto done;
             }
             free(include_file);
             include_file = 0;
@@ -246,6 +406,20 @@ int flatcc_parse_file(flatcc_context_t ctx, const char *filename)
         ptr_set_insert_item(&P->schema.visible_schema, &P->schema, ht_keep);
         ret = fb_build_schema(P);
     }
+
+    /*
+     * We choose to only generate optional .depends files for root level
+     * files. These will contain all nested files regardless of
+     * recursive file generation flags.
+     */
+    if (P->opts.gen_dep && is_root) {
+        ret = __flatcc_gen_depends_file(P);
+    }
+
+done:
+    checkfree(buf);
+    checkfree(path);
+    checkfree(include_file);
     return ret;
 }
 
@@ -271,7 +445,9 @@ void *flatcc_generate_binary_schema(flatcc_context_t ctx, size_t *size)
 int flatcc_generate_files(flatcc_context_t ctx)
 {
     fb_parser_t *P = ctx, *P_leaf;
+    fb_output_t *out, output;
     int ret = 0;
+    out = &output;
 
     if (!P || P->failed) {
         return -1;
@@ -290,23 +466,30 @@ int flatcc_generate_files(flatcc_context_t ctx)
         }
     }
 #endif
-    /* This does not require a parse first. */
-    if (fb_codegen_common_c(&P->opts)) {
+
+    if (fb_init_output_c(out, &P->opts)) {
         return -1;
+    }
+    /* This does not require a parse first. */
+    if (!P->opts.gen_append && (ret = fb_codegen_common_c(out))) {
+        goto done;
     }
     /* If no file parsed - just common files if at all. */
     if (!P->has_schema) {
-        return 0;
+        goto done;
     }
     if (!P->opts.cgen_recursive) {
-        return fb_codegen_c(&P->opts, &P->schema);
+        ret = fb_codegen_c(out, &P->schema);
+        goto done;
     }
-    /* Make sure stdout output is generated in the right order. */
+    /* Make sure stdout and outfile output is generated in the right order. */
     P = P_leaf;
     while (!ret && P) {
-        ret = P->failed || fb_codegen_c(&P->opts, &P->schema);
+        ret = P->failed || fb_codegen_c(out, &P->schema);
         P = P->inverse_dependencies;
     }
+done:
+    fb_end_output_c(out);
     return ret;
 }