+++ test/json_test/test_json_printer.c
@@ -18,15 +18,16 @@
 /* A helper to simplify creating buffers vectors from C-arrays. */
 #define c_vec_len(V) (sizeof(V)/sizeof((V)[0]))
 
+const char *filename = "monsterdata_test.mon";
+const char *golden_filename = "monsterdata_test.golden";
+const char *target_filename = "monsterdata_test.json.txt";
+
 int test_print()
 {
     int ret = 0;
     const char *buf = 0;
     const char *golden = 0;
     const char *target = 0;
-    const char *filename = "monsterdata_test.mon";
-    const char *golden_filename = "monsterdata_test.golden";
-    const char *target_filename = "monsterdata_test.json.txt";
     size_t size = 0, golden_size = 0, target_size = 0;
     flatcc_json_printer_t ctx_obj, *ctx;
     FILE *fp = 0;
@@ -91,12 +92,22 @@ fail:
     goto done;
 }
 
+/* We take arguments so output file can be generated in build directory without copying sources. */
+#define usage \
+"wrong number of arguments:\n" \
+"usage: <program> [<input-filename> <reference-filename> <output-filename>]\n"
+
 int main(int argc, const char *argv[])
 {
-    (void)argc;
-    (void)argv;
-
     fprintf(stderr, "running json print test\n");
+    if (argc != 1 && argc != 4) {
+        fprintf(stderr, usage);
+        exit(1);
+    }
+    if (argc == 4) {
+        filename = argv[1];
+        golden_filename = argv[2];
+        target_filename = argv[3];
+    }
     return test_print();
 }
-